Add include guard :guard:
This prevents subdirectories that use an older version of the script
from overriding the (newer) version that parents include.

We specifically do not use `include_guard` since it probably (did not
actually verify this but the documentation seem to suggest it is so)
will not work with multiple instances of dep within a hierarchy, as
one would get with e.g. submodules in git.